Move sign_encode test from entry into encoder
